refactor: migrate from PostgreSQL to SQLite with better-sqlite3
Replace PostgreSQL + Electric SQL with a single SQLite database using
better-sqlite3 for simpler deployment and zero external dependencies.
Consolidate 7 migration files into one, remove shapes route and Electric
client, simplify Docker from 3 services to 2, and auto-generate JWT
secrets when env var is unset.

+++ server/src/routes/activity.ts
@@ -31,7 +31,7 @@ router.get('/:locationId/activity', requireLocationMember('locationId'), async (
 
     // Get total count
     const countResult = await query(
-      `SELECT COUNT(*)::int AS count FROM activity_log al ${whereClause}`,
+      `SELECT COUNT(*) AS count FROM activity_log al ${whereClause}`,
       params
     );
     const count = countResult.rows[0].count;
